+++ src/onegov.feriennet/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Rely on latest onegov.org release.
+  [href]
+
 0.6.0 (2017-03-15)
 ~~~~~~~~~~~~~~~~~~~